fastapi_staging/requirements.txt
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.3.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.13.4
billiard==4.2.1
black==25.1.0
bleach==6.2.0
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.6
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
eventlet==0.40.2
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.2
gevent==25.5.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kombu==5.5.4
limits==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pyzmq==27.0.1
redis==5.2.1
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
ruff==0.12.4
setuptools==80.9.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.47.2
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.17.2
yarg==0.1.9
zope.event==5.1.1
zope.interface==7.2
